requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
prisma==0.11.0
jinja2==3.1.3
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.1.0